Output in long long to prevent printing in exponential form.

+++ theatresq.cpp
@@ -1,12 +1,14 @@
 #include<iostream>
 #include<math.h>
+#include<iomanip>
+
 using namespace std;
 int main()
 {
 	int n, m, a;
 	cin >> n >> m >> a;
 
-	cout << ceil(n / (double)a) * ceil(m / (double)a) << endl;
+	cout << (long long)(ceil(n / (double)a) * ceil(m / (double)a)) << endl;
 
 	return 0;
 }
